+++ src/cli.py
@@ -6,7 +6,7 @@ import os
 from src.database import db
 
 cwd = os.path.dirname(os.path.abspath(__file__))
-DB_ENGINE = db.create_engine('sqlite:////{}/test.db'.format(cwd))
+DB_ENGINE = db.create_engine("sqlite:////{}/test.db".format(cwd))
 
 
 def register_cli_commands(app):
@@ -15,7 +15,8 @@ def register_cli_commands(app):
     Args:
         app (flask app object): Flask application object
     """
-    @app.cli.command('initdb')
+
+    @app.cli.command("initdb")
     def init_db_command():
         """Initialize the database. """
         db.init_app(app)
@@ -23,11 +24,11 @@ def register_cli_commands(app):
         connection = DB_ENGINE.connect()
         connection.close()
         db.create_all()
-        app.logger.info('Database initialized')
+        app.logger.info("Database initialized")
 
-    @app.cli.command('dropdb')
+    @app.cli.command("dropdb")
     def drop_db_command():
         """Drop the database. """
         db.init_app(app)
         db.drop_all()
-        app.logger.info('Database dropped')
+        app.logger.info("Database dropped")