+++ src/database/__init__.py
@@ -5,5 +5,6 @@ db = SQLAlchemy()
 
 def reset_database():
     from src.database.models import Resources, Bookings
+
     db.drop_all()
     db.create_all()